requirements.txt
pydantic
gradio
numpy
pillow
opencv-python
pyyaml
